# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/operator/pgsqloperator": {
            "post": {
                "description": "CreateOperator deploy pgsqloperator to the kubernetes cluster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operator"
                ],
                "summary": "CreateOperator deploy pgsqloperator to the kubernetes cluster",
                "parameters": [
                    {
                        "description": "Deploy pgsqloperator params",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pgsql.CreateOpeartorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"OK\",\"data\":{\"Namespace\":\"defalt\", \"DeploymentName\": \"postgres-operator\"}}",
                        "schema": {
                            "$ref": "#/definitions/pgsql.CreateOperatorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "DeleteOperator delete pgsqloperator from the kubernetes cluster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operator"
                ],
                "summary": "DeleteOperator delete pgsqloperator from the kubernetes cluster",
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"message\":\"OK\",\"data\":{\"\"}}",
                        "schema": {
                            "$ref": "#/definitions/tool.Response"
                        }
                    }
                }
            }
        },
        "/operator/redisoperator": {
            "post": {
                "description": "CreateOperator 安装 redis operator.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operator"
                ],
                "summary": "CreateOperator 安装 redis operator.",
                "parameters": [
                    {
                        "description": "安装Redis operator所需参数.",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/redis.CreateOperatorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":0,\"message\":\"OK\",\"data\":{\"\"}}",
                        "schema": {
                            "$ref": "#/definitions/tool.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "从Kubernetes集群中删除redis operator组件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operator"
                ],
                "summary": "从Kubernetes集群中删除redis operator组件",
                "parameters": [
                    {
                        "description": "删除Redis Operator组件时所需的参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/redis.DeleteOperatorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"message\":\"OK\",\"data\":{\"\"}}",
                        "schema": {
                            "$ref": "#/definitions/tool.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pgsql.CreateOpeartorRequest": {
            "type": "object",
            "properties": {
                "ClusterRoleBindingName": {
                    "type": "string"
                },
                "ClusterRoleName": {
                    "type": "string"
                },
                "ConfigMapName": {
                    "type": "string"
                },
                "DeploymentName": {
                    "type": "string"
                },
                "Namespace": {
                    "type": "string"
                },
                "PostgresqlOperatorImage": {
                    "type": "string"
                },
                "serviceAccountName": {
                    "type": "string"
                }
            }
        },
        "pgsql.CreateOperatorResponse": {
            "type": "object",
            "properties": {
                "ClusterRoleBindingName": {
                    "type": "string"
                },
                "ClusterRoleName": {
                    "type": "string"
                },
                "ConfigMapName": {
                    "type": "string"
                },
                "DeploymentName": {
                    "type": "string"
                },
                "Namespace": {
                    "type": "string"
                },
                "serviceAccountName": {
                    "type": "string"
                }
            }
        },
...
        "redis.CreateOperatorRequest": {
            "type": "object",
            "properties": {
                "namespace": {
                    "type": "string"
                }
            }
        },
...
        "redis.DeleteOperatorRequest": {
            "type": "object",
            "properties": {
                "namespace": {
                    "type": "string"
                }
            }
        },
...
        "tool.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}